Refactor GUI into separate modules
