+++ dialogs/__init__.py
@@ -1,0 +1,5 @@
+from .batch_mapping_dialog import BatchMappingDialog
+from .mapping_dialog import MappingDialog
+from .drop_area import DropArea
+
+__all__ = ["BatchMappingDialog", "MappingDialog", "DropArea"]
