+++ main.py
@@ -1,0 +1,4 @@
+from main_window import main
+
+if __name__ == "__main__":
+    main()
