+++ main_window.py
@@ -935,5 +935,3 @@ def main():
     sys.exit(app.exec())
 
 
-if __name__ == "__main__":
-    main()
